Moved Role and User to Entities Folder
