+++ Orso.Arpa.Application/Auth/ChangePassword.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Identity;
 using Orso.Arpa.Application.Errors;
 using Orso.Arpa.Application.Interfaces;
 using Orso.Arpa.Application.Validators;
-using Orso.Arpa.Domain;
+using Orso.Arpa.Domain.Entities;
 
 namespace Orso.Arpa.Application.Auth
 {
