+++ Orso.Arpa.Application/Auth/Login.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.Identity;
 using Orso.Arpa.Application.Auth.Dtos;
 using Orso.Arpa.Application.Errors;
 using Orso.Arpa.Application.Interfaces;
-using Orso.Arpa.Domain;
+using Orso.Arpa.Domain.Entities;
 
 namespace Orso.Arpa.Application.Auth
 {
