+++ Orso.Arpa.Application/Auth/SetRole.cs
@@ -7,6 +7,7 @@ using MediatR;
 using Microsoft.AspNetCore.Identity;
 using Orso.Arpa.Application.Errors;
 using Orso.Arpa.Domain;
+using Orso.Arpa.Domain.Entities;
 
 namespace Orso.Arpa.Application.Auth
 {
