+++ Orso.Arpa.Application/Interfaces/IJwtGenerator.cs
@@ -1,5 +1,5 @@
 using System.Threading.Tasks;
-using Orso.Arpa.Domain;
+using Orso.Arpa.Domain.Entities;
 
 namespace Orso.Arpa.Application.Interfaces
 {
