+++ Orso.Arpa.Application/Users/List.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using MediatR;
 using Microsoft.AspNetCore.Identity;
 using Orso.Arpa.Application.Users.Dtos;
-using Orso.Arpa.Domain;
+using Orso.Arpa.Domain.Entities;
 
 namespace Orso.Arpa.Application.Users
 {
