+++ Orso.Arpa.Domain/Entities/Role.cs
@@ -1,7 +1,7 @@
 using System;
 using Microsoft.AspNetCore.Identity;
 
-namespace Orso.Arpa.Domain
+namespace Orso.Arpa.Domain.Entities
 {
     public class Role : IdentityRole<Guid>
     {
