+++ Orso.Arpa.Domain/Entities/User.cs
@@ -1,7 +1,7 @@
 using System;
 using Microsoft.AspNetCore.Identity;
 
-namespace Orso.Arpa.Domain
+namespace Orso.Arpa.Domain.Entities
 {
     public class User : IdentityUser<Guid>
     {
