+++ Orso.Arpa.Domain/SeedData/RoleSeedData.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Orso.Arpa.Domain.Entities;
 
 namespace Orso.Arpa.Domain.Seed
 {
