+++ Orso.Arpa.Infrastructure/Security/JwtGenerator.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Configuration;
 using Microsoft.IdentityModel.Tokens;
 using Orso.Arpa.Application.Interfaces;
-using Orso.Arpa.Domain;
+using Orso.Arpa.Domain.Entities;
 
 namespace Orso.Arpa.Infrastructure.Security
 {
