+++ Orso.Arpa.Infrastructure/Security/UserAccessor.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Orso.Arpa.Application.Errors;
 using Orso.Arpa.Application.Interfaces;
-using Orso.Arpa.Domain;
+using Orso.Arpa.Domain.Entities;
 
 namespace Orso.Arpa.Infrastructure.Security
 {
