+++ Orso.Arpa.Persistence/ArpaContext.cs
@@ -1,7 +1,7 @@
 using System;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
-using Orso.Arpa.Domain;
+using Orso.Arpa.Domain.Entities;
 using Orso.Arpa.Persistence.Configurations;
 
 namespace Orso.Arpa.Persistence
