+++ Tests/Orso.Arpa.Api.Tests/IntegrationTests/Shared/AuthenticatedTestRequestMiddleware.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Newtonsoft.Json;
 using Orso.Arpa.Application.Interfaces;
-using Orso.Arpa.Domain;
+using Orso.Arpa.Domain.Entities;
 
 namespace Orso.Arpa.Api.Tests.IntegrationTests.Shared
 {
