+++ Tests/Orso.Arpa.Api.Tests/IntegrationTests/UsersControllerTests.cs
@@ -7,6 +7,7 @@ using NUnit.Framework;
 using Orso.Arpa.Api.Tests.IntegrationTests.Shared;
 using Orso.Arpa.Application.Users;
 using Orso.Arpa.Application.Users.Dtos;
+using Orso.Arpa.Domain.Entities;
 using Orso.Arpa.Tests.Shared.DtoTestData;
 using Orso.Arpa.Tests.Shared.TestSeedData;
 
@@ -19,7 +20,7 @@ namespace Orso.Arpa.Api.Tests.IntegrationTests
         public async Task Should_Delete_User()
         {
             // Arrange
-            Domain.User user = UserSeedData.Orsoadmin;
+            User user = UserSeedData.Orsoadmin;
             var command = new Delete.Command(user.UserName);
 
             // Act
@@ -36,7 +37,7 @@ namespace Orso.Arpa.Api.Tests.IntegrationTests
         public async Task Should_Not_Delete_Deleted_User()
         {
             // Arrange
-            Domain.User user = UserSeedData.DeletedUser;
+            User user = UserSeedData.DeletedUser;
             var command = new Delete.Command(user.UserName);
 
             // Act
@@ -72,7 +73,7 @@ namespace Orso.Arpa.Api.Tests.IntegrationTests
         public async Task Should_Not_Delete_User_If_Current_User_Is_Not_Orsoadmin()
         {
             // Arrange
-            Domain.User user = UserSeedData.Orsoadmin;
+            User user = UserSeedData.Orsoadmin;
 
             // Act
             HttpResponseMessage responseMessage = await _authenticatedServer
