+++ Tests/Orso.Arpa.Application.Tests/AuthTests/QueryHandlerTests/LoginHandlerTests.cs
@@ -7,7 +7,7 @@ using NUnit.Framework;
 using Orso.Arpa.Application.Auth;
 using Orso.Arpa.Application.Auth.Dtos;
 using Orso.Arpa.Application.Interfaces;
-using Orso.Arpa.Domain;
+using Orso.Arpa.Domain.Entities;
 using Orso.Arpa.Tests.Shared.Identity;
 using Orso.Arpa.Tests.Shared.TestSeedData;
