+++ Tests/Orso.Arpa.Application.Tests/AuthTests/ValidationTests/ChangePasswordCommandValidatorTests.cs
@@ -4,7 +4,7 @@ using NSubstitute;
 using NUnit.Framework;
 using Orso.Arpa.Application.Auth;
 using Orso.Arpa.Application.Interfaces;
-using Orso.Arpa.Domain;
+using Orso.Arpa.Domain.Entities;
 using Orso.Arpa.Tests.Shared.Identity;
 using Orso.Arpa.Tests.Shared.TestSeedData;
