+++ Tests/Orso.Arpa.Application.Tests/AuthTests/ValidationTests/RegisterCommandValidatorTests.cs
@@ -2,7 +2,7 @@ using FluentValidation.TestHelper;
 using Microsoft.AspNetCore.Identity;
 using NUnit.Framework;
 using Orso.Arpa.Application.Auth;
-using Orso.Arpa.Domain;
+using Orso.Arpa.Domain.Entities;
 using Orso.Arpa.Tests.Shared.Identity;
 using Orso.Arpa.Tests.Shared.TestSeedData;
