+++ Tests/Orso.Arpa.Infrastructure.Tests/SecurityTests/UserAccessorTests.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Identity;
 using NSubstitute;
 using NUnit.Framework;
 using Orso.Arpa.Application.Errors;
-using Orso.Arpa.Domain;
+using Orso.Arpa.Domain.Entities;
 using Orso.Arpa.Infrastructure.Security;
 using Orso.Arpa.Tests.Shared.Identity;
 using Orso.Arpa.Tests.Shared.TestSeedData;
